tweaked examples
