+++ packages/xchain-thorchain-amm/examples/xchainjs-test/doSwap.ts
@@ -1,5 +1,5 @@
 import { Network } from '@xchainjs/xchain-client'
-import { CryptoAmount, EstimateSwapParams, Midgard, SwapEstimate, ThorchainAMM } from '@xchainjs/xchain-thorchain-amm'
+import { CryptoAmount, Midgard, SwapEstimate, ThorchainAMM, Wallet } from '@xchainjs/xchain-thorchain-amm'
 import { assetAmount, assetFromString, assetToBase } from '@xchainjs/xchain-util'
 import BigNumber from 'bignumber.js'
 
@@ -23,37 +23,39 @@ function print(estimate: SwapEstimate, input: CryptoAmount) {
 }
 
 /**
- * Estimate swap function
- * Returns estimate swap object
+ * From asset to asset with no Affiliate address on testnet
  */
-const estimateSwap = async () => {
+const doSingleSwap = async (tcAmm: ThorchainAMM, wallet: Wallet) => {
   try {
-    const network = process.argv[2] as Network
-    const amount = process.argv[3]
-    const fromAsset = assetFromString(`${process.argv[4]}`)
-    const toAsset = assetFromString(`${process.argv[5]}`)
-    const midgard = new Midgard(network)
-    const thorchainAmm = new ThorchainAMM(midgard)
+    const amount = process.argv[4]
+    const fromAsset = assetFromString(`${process.argv[5]}`)
+    const toAsset = assetFromString(`${process.argv[6]}`)
 
-    const swapParams: EstimateSwapParams = {
+    const swapParams = {
       input: new CryptoAmount(assetToBase(assetAmount(amount)), fromAsset),
       destinationAsset: toAsset,
-      // affiliateFeePercent: 0.003, //optional
       slipLimit: new BigNumber('0.03'), //optional
     }
-    const estimate = await thorchainAmm.estimateSwap(swapParams)
-    print(estimate, swapParams.input)
-    const estimateInFromAsset = await thorchainAmm.getFeesIn(estimate.totalFees, fromAsset)
-    estimate.totalFees = estimateInFromAsset
-    print(estimate, swapParams.input)
-  } catch (e) {
-    console.error(e)
+
+    const outPutCanSwap = await tcAmm.estimateSwap(swapParams)
+    print(outPutCanSwap, swapParams.input)
+    if (outPutCanSwap.canSwap) {
+      const output = await tcAmm.doSwap(wallet, swapParams, wallet.clients[toAsset.chain].getAddress())
+      console.log(`Tx hash: ${output.hash},\n Tx url: ${output.url}\n WaitTime: ${output.waitTimeSeconds}`)
+    }
+  } catch (error) {
+    console.error(error)
   }
 }
 
-// Call the function from main()
 const main = async () => {
-  await estimateSwap()
+  const seed = process.argv[2]
+  const network = process.argv[3] as Network
+  const midgard = new Midgard(network)
+  const thorchainAmm = new ThorchainAMM(midgard)
+  const wallet = new Wallet(network, seed)
+  console.log(`\ Swap on ${network} :)\n`)
+  await doSingleSwap(thorchainAmm, wallet)
 }
 
 main()
