+++ packages/xchain-thorchain-amm/examples/xchainjs-test/listPools.ts
@@ -1,0 +1,13 @@
+import { Network } from '@xchainjs/xchain-client'
+import { Midgard } from '@xchainjs/xchain-thorchain-amm'
+
+const main = async () => {
+  const network = process.argv[2] as Network
+
+  const midgard = new Midgard(network)
+  console.table(await midgard.getPools())
+}
+
+main()
+  .then(() => process.exit(0))
+  .catch((err) => console.error(err))
