+++ packages/xchain-thorchain-amm/examples/xchainjs-test/package.json
@@ -2,9 +2,9 @@
   "name": "xchainjs-test",
   "version": "1.0.0",
   "scripts": {
-    "estimateSwap": "ts-node estimateSwap.ts",
-    "doSingleSwap": "ts-node doSwap-Single.ts",
-    "doDoubleSwap": "ts-node doSwap-Double.ts"
+    "listPools": "npx ts-node listPools.ts",
+    "estimateSwap": "npx ts-node estimateSwap.ts",
+    "doSwap": "npx ts-node doSwap.ts"
   },
   "description": "Test folder for xchainjs thorchain-amm package",
   "main": "index.js",
